% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_globals.R
\name{check_globals}
\alias{check_globals}
\title{Check if functions use global variables}
\usage{
check_globals(env = .GlobalEnv)
}
\arguments{
\item{env}{a specified environment to check, default \code{.GlobalEnv}}
}
\value{
a \code{data.frame} with columns \code{Function}, \code{Env}, and \code{HasGlobals}
}
\description{
Checks all the defined functions in a specified environment whether they reference global variables, using \code{codetools::findGlobals()}.

Ideally most functions would be pure, but quick one-off functions might be expected to reference global variables.
}
\examples{
mk::check_globals()
}
