% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{dt}
\alias{dt}
\title{Wrapper for DT::datatable()}
\usage{
dt(
  df,
  title = "",
  subtitle = c(""),
  page.length = 10,
  plain = FALSE,
  row.names = FALSE,
  width = "95\%",
  cols.round.1 = NULL,
  cols.round.2 = NULL,
  cols.round.3 = NULL,
  cols.round.4 = NULL,
  cols.nowrap = NULL,
  cols.pval = NULL,
  cols.pos.neg = NULL,
  cols.fac.color = NULL,
  lighten = 0.35
)
}
\arguments{
\item{df}{a dataframe}

\item{title, subtitle}{title and subtitle(s)}

\item{page.length}{length of each page}

\item{plain}{FALSE (default) = typical datatable with controls, search, sorting, etc.

TRUE = more simplified, less controls, sets \code{dom = "tr"} within \code{DT::datatable(options = ...)}}

\item{row.names}{Whether to show row.names() of data.frame, default = FALSE}

\item{width}{width e.g., "95\%" or "500px" or 500, less than 95\% helps with unnecessary scrollbars}

\item{cols.round.1}{colnames/indices in which to round to 1 decimal place}

\item{cols.round.2}{colnames/indices in which to round to 2 decimal places}

\item{cols.round.3}{colnames/indices in which to round to 3 decimal places}

\item{cols.round.4}{colnames/indices in which to round to 4 decimal places}

\item{cols.nowrap}{colnames/indices where content will not be wrapped}

\item{cols.pval}{column names/indices where cells with numbers [0, 0.05) will be shaded yellow and [0.05, 0.15) shaded orange}

\item{cols.pos.neg}{column names/indices where cells with numbers will be shaded green/red for positive/negative values}

\item{cols.fac.color}{default NULL for no coloring, or one of:
\itemize{
\item character or numeric vector = provide colnames/indices for character columns
\item list where list names = column names and list items = \if{html}{\out{<chr>}} hex codes / colors for levels in the column, list items may or may not be named characters where names() refer to the levels
}}

\item{lighten}{lighten the provided colors to argument "cols.fac.color" by a factor where 1 = original color and 0 = white, default = 0.35, NULL for no lightening}
}
\value{
html widget
}
\description{
Takes a dataframe and makes an interactive html table
}
\details{
Re: arguments \code{cols.fac.color} and \code{lighten}: DT renders with CSS values of colors, so "green" may be different than R green, and !is.null(lighten) forces hex codes instead of color "green", so lighten = NULL and lighten = 1, may yield different results
}
\examples{
# dataframe to summarize
df <- lapply(split(iris, iris$Species), \(df) {
  df[sample(nrow(df), 5), ]
}) |> do.call(rbind, args = _)
df$pval <- seq(0.01, 0.15, 0.01)

# Various options
mk::dt(
  df = df,
  title = "A title",
  subtitle = c("Line one of the subtitle", "and line two"),
  page.length = 15,
  plain = TRUE,
  row.names = FALSE,
  # Specify columns by indices or names
  cols.nowrap = colnames(df),
  cols.round.2 = "Petal.Width",
  cols.round.3 = 1,
  cols.round.4 = c("Sepal.Width", "Petal.Length", "pval"),
  cols.pval = "pval",
  cols.fac.color =  "Species"
)

# Explicit colors
mk::dt(
  df = df[, c("Species", "pval")],
  plain = TRUE,
  page.length = 15,
  width = 200,
  cols.round.4 = 2,
  cols.fac.color = list(
    Species = c("#EA4F0D", "#29EFA2", "#4490FE")
  )
)
}
