% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_curve.R
\name{enrich_curve}
\alias{enrich_curve}
\title{Enrichment curve for GSEA results}
\usage{
enrich_curve(
  id,
  gsea.res,
  title = id,
  subtitle = NULL,
  caption = NULL,
  plot.stats = TRUE,
  sig = c("qvalue", "p.adjust", "pvalue"),
  base.size = 16,
  ...
)
}
\arguments{
\item{id}{one gene set id, from \code{ID} column of gsea.res}

\item{gsea.res}{GSEA result object of class "gseaResult", e.g., output of \code{clusterProfiler::GSEA()}}

\item{title, subtitle, caption}{ggplot titles, etc.}

\item{plot.stats}{Whether to add NES and test statistic to plot, default TRUE}

\item{sig}{For plot.stats = TRUE, which significance to use? default "qvalue"}

\item{base.size}{pass argument to mk::thm() to change default sizes of text elements in plot, default NULL to take the default value in mk::thm()}

\item{...}{Arguments merged to enrichplot::gseaplot()}
}
\value{
A ggplot object
}
\description{
Take "gseaResult" and return enrichment curve plot
}
\examples{
\dontrun{
mk::enrich_curve(gene.set.id, gsea.res) +
 mk::size(5, 2.5)
mk::enrich_curve(gene.set.id, gsea.res, plot.stats = FALSE) +
 mk::size(5, 2.5)
}
}
