% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colors.R
\name{gg_colors}
\alias{gg_colors}
\title{Generate ggplot2 default color palette}
\usage{
gg_colors(n)
}
\arguments{
\item{n}{number of colors to generate}
}
\value{
a character vector of hex values
}
\description{
Pull values equally spaced from the color wheel
}
\details{
Code adapted from this Stack Overflow question: https://stackoverflow.com/questions/8197559
}
\examples{
mk::gg_colors(8)
}
