% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sci.R
\name{label_sci}
\alias{label_sci}
\title{Axis labels for scientific notation}
\usage{
label_sci(x)
}
\arguments{
\item{x}{a numeric vector of breaks to format}
}
\value{
a vector of \code{expression()} results
}
\description{
Supply this function to the \code{labels} argument of ggplot scale function, e.g., \code{ggplot2::scale_y_continuous()}
}
\details{
This function first forces scientific notation with \code{scales::label_scientific()} then beautifies it a bit with e.g., 5x10^2 instead of 5e+02
}
\examples{
\dontrun{
p + ggplot2::scale_y_continuous(labels = mk::label_sci)
}

}
