% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten_colors.R
\name{lighten_colors}
\alias{lighten_colors}
\alias{darken_colors}
\title{Lighten / Darken a color palette}
\usage{
lighten_colors(colors, factor)

darken_colors(colors, factor)
}
\arguments{
\item{colors}{\if{html}{\out{<chr>}} color names or hex codes}

\item{factor}{factor for lightening or darkening, where:
\itemize{
\item for mk::lighten_colors(), 1 = original color and 0 = white
\item for mk::darken_colors(), 1 = original color and 0 = black
}}
}
\value{
a character vector of hex codes
}
\description{
Lighten or darken a color palette with mk::lighten_colors() or mk::darken_colors(). Inspired by / adapted from https://gist.github.com/Jfortin1/72ef064469d1703c6b30, 2024-10-13
}
\examples{
mk::lighten_colors(c("red", "green", "blue"), 0.65)
mk::darken_colors(mk::gg_colors(8), 0.75)

}
