% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk.R
\name{mk.tiprate}
\alias{mk.tiprate}
\title{Compute tip rates using parsimony}
\usage{
mk.tiprate(x, phy)
}
\arguments{
\item{x}{A named vector of character state data having type \code{factor} 
or \code{integer}.}

\item{phy}{An object of class \code{tree}.}
}
\value{
A vector of tip rates
}
\description{
The average tip rate approximates the maximum likelihood estimate of the
rate parameter of a fully symmetric Markov model.
}
\examples{
data(squamatatree)
data(squamatareprod)
phy = read.newick(text=squamatatree)
x = mk.tiprate(squamatareprod, phy)
# average tiprate
mean(x) # 0.002310957
# maximum likelihood rate estimate
optimize(mk(squamatareprod, phy), c(0,1), maximum=TRUE)$maximum # 0.001506275
}
