% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_sizes.R
\name{object_sizes}
\alias{object_sizes}
\title{Get sizes of objects in memory}
\usage{
object_sizes(sort = "desc")
}
\arguments{
\item{sort}{"desc" (default) to sort by descending size or NULL for default order from \code{ls()}}
}
\value{
a dataframe of object sizes in Gb or Mb for all items retrieved from ls(envir = .GlobalEnv).

if nothing found from ls(), return(NULL)
}
\description{
Get sizes of objects in memory
}
\details{
This reports sizes in GB or MB in character vector output, not suitable from this output for further calculations.
}
\examples{
mk::object_sizes()

}
