% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot color palettes}
\usage{
plot_colors(
  colors,
  ncol = 8,
  text.size = 0,
  pal.text.size = 5,
  title.loc = c("right", "above"),
  pal.spacing = NULL,
  right.margin = NULL,
  linewidth = 0.5,
  cell.aspect = 0.5,
  dev = FALSE
)
}
\arguments{
\item{colors}{colors to plot, \if{html}{\out{<chr>}} vector or list of \if{html}{\out{<chr>}} vectors (named or unnamed)}

\item{ncol}{number of colors per row plotted, default 8}

\item{text.size}{text size of color/hex values inside cells, default 0}

\item{pal.text.size}{text size of labels per each color palette, default 5}

\item{title.loc}{location of palette titles, one of "right" (default) or "above"}

\item{pal.spacing}{spacing between palettes (in units of cell height), default NULL is 0.25 for title.loc = "right" or 1.25 for title.loc = "above", otherwise provide numeric}

\item{right.margin}{margin to the right, e.g., for title.loc = "right". default NULL is 0 for title.loc = "above", or attempts to guess for title.loc = "right". Otherwise, provide \if{html}{\out{<numeric>}} in inches}

\item{linewidth}{border linewidth, default 0.5}

\item{cell.aspect}{aspect ratio per color cell (height / width), \if{html}{\out{<numeric>}}. default 0.5, NULL for no aspect.ratio applied.}

\item{dev}{Don't plot, just return the dataframe used for plotting, default FALSE}
}
\value{
a ggplot object
}
\description{
Provide a vector of colors, or named list of multiple palettes, and return a plot
}
\details{
Each color gets a shape from ggplot2::geom_rect which is spaced at x-y coordinates
}
\examples{
# Plot a single palette
mk::plot_colors(mk::gg_colors(8))

# Customize the palette title shown
mk::plot_colors(list(`Default ggplot2 colors` = mk::gg_colors(8)))

# Plot in multiple rows and multiple palettes
# ncol default is 8
mk::plot_colors(
  colors = list(
    `viridis::turbo(12)` = viridis::turbo(12),
    `RColorBrewer::brewer.pal(8, "Set2")` = RColorBrewer::brewer.pal(8, "Set2")
  ),
  ncol = 4
)

# Put palette title above each palette instead of to the right
# Show color values / hex codes inside cells
mk::plot_colors(
  colors = list(
    `RColorBrewer::brewer.pal(9, "Reds")`      = RColorBrewer::brewer.pal(9, "Reds"),
    `RColorBrewer::brewer.pal(9, "Blues")`     = RColorBrewer::brewer.pal(9, "Blues"),
    `RColorBrewer::brewer.pal(11, "Spectral")` = RColorBrewer::brewer.pal(11, "Spectral")
  ),
  ncol = 11,
  text.size = 4,
  title.loc = "above"
)
}
