% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\alias{sp}
\title{Plot save wrapper function}
\usage{
save_plot(
  p,
  width = 7,
  height = 5,
  file = "Rplot.png",
  type = NULL,
  units = "in",
  device = NULL,
  bg = "transparent",
  dpi = 300,
  auto.dpi = 300,
  print.capabilities = FALSE,
  dpi.CairoPDF = "auto",
  dpi.CairoSVG = "auto",
  compression = "lzw",
  quality = 75,
  silent = FALSE,
  ...
)

sp(
  p,
  width = 7,
  height = 5,
  file = "Rplot.png",
  type = NULL,
  units = "in",
  device = NULL,
  bg = "transparent",
  dpi = 300,
  auto.dpi = 300,
  print.capabilities = FALSE,
  dpi.CairoPDF = "auto",
  dpi.CairoSVG = "auto",
  compression = "lzw",
  quality = 75,
  silent = FALSE,
  ...
)
}
\arguments{
\item{p}{plot object invoked with plot()}

\item{width}{width (inches by default), default = 7}

\item{height}{height (inches by default), default = 5}

\item{file}{path to file WITH extension, default = "Rplot.png"}

\item{type}{image type output. default NULL = infer from \code{file} extension. Or one of: bmp, jpeg, pdf, png, svg, tiff}

\item{units}{units for width and height, ONLY current option is inches (default units = "in")}

\item{device}{default NULL = choose device based on those available. Use \code{print.capabilities = TRUE} to see valid options. Depending on installation, one of: bmp, Cairo, cairo_pdf, jpeg, png, svg, tiff. Cairo device (Cairo package) supports multiple types}

\item{bg}{background color \if{html}{\out{<chr>}}, default "transparent", or a R color or hex value \if{html}{\out{<chr>}}}

\item{dpi}{\if{html}{\out{<numeric>}} dots per inch, default 300}

\item{auto.dpi}{\if{html}{\out{<numeric>}} default 300. If argument \code{dpi} not supplied, set dpi such that an image scaled to 13.5 inch width would be approx this many dpi. Ignored if not using a rasterized format (png, jpeg, tiff, bmp). NULL = turn off this functionality.}

\item{print.capabilities}{TRUE = Don't plot, just return dataframe of supported and installed types/devices on this system. default = FALSE.}

\item{dpi.CairoPDF, dpi.CairoSVG}{dpi arguments when device = "Cairo" and type = "pdf" or "svg". default = "auto". See details.}

\item{compression}{compression used for device = "tiff". default = "lzw". Otherwise file types are very large for compression = "none".}

\item{quality}{quality used for device = "jpeg". default = 75 \if{html}{\out{<numeric>}}.}

\item{silent}{silence messages from this function. Does not silences messages, warnings, or errors from \code{plot()}.}

\item{...}{Arguments passed to selected device}
}
\value{
invisible(NULL)
}
\description{
Quickly save plots
}
\details{
...
}
\examples{
\dontrun{
# See capabilities available
save_plot(print.capabilities = TRUE)

# sp() is an alias for save_plot()
sp(print.capabilties = TRUE)

# Quick plot to default file = "Rplot.png"
sp(p, 5, 3)

# Type is inferred by file extension
sp(p, 5, 3, file = "Rplot.pdf")

# Explicitly define dpi with dpi argument to
  # avoid auto.dpi argument behavior
save_plot(
  p      = p,
  width  = 7,
  height = 5,
  units  = "in",
  file   = "myplot.tiff",
  dpi    = 800
)

# If multiple devices capable of a file type, can
  # explicitly define device, e.g., cairo_pdf instead
  # of default Cairo::Cairo(type = "pdf")
sp(p, 5, 3, file = "Rplot.pdf", device = "cairo_pdf")

# Save with an interesting file extension
sp(p, 5, 3, file = "Rplot.hello", type = "png")

}
}
