% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_check.R
\name{se_check}
\alias{se_check}
\title{Check SE object structure}
\usage{
se_check(se, silent = FALSE, strict = FALSE)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{silent}{suppress messages, default FALSE}

\item{strict}{more opinionated checks, default FALSE}
}
\value{
TRUE/FALSE
}
\description{
Check agreement in dimnames of the assays and the rowData()/colData()
}
\details{
This function checks:
\itemize{
\item that \code{row.names()} of all assays identical to \code{row.names(rowData(se))}
\item that \code{colnames()} of all assays identical to \code{row.names(colData(se))}
}

For strict = TRUE
\itemize{
\item that \code{as.character(rowData(se)[[1]])} identical to \code{row.names(rowData(se))}
\item that \code{as.character(colData(se)[[1]])} identical to \code{row.names(colData(se))}
\item that
}
}
\examples{
\dontrun{
mk::se_check(se)
mk::se_check(se, silent = TRUE)
mk::se_check(se, strict = TRUE)
}
}
