% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_label_pair.R
\name{so_label_pair}
\alias{so_label_pair}
\title{Seurat number-label pairs for DimPlots}
\usage{
so_label_pair(so, ident, start.zero = TRUE)
}
\arguments{
\item{so}{a Seurat object}

\item{ident}{an ident from which to create a number-label pair}

\item{start.zero}{TRUE (default) = start numbering at zero

FALSE = start numbering at one}
}
\value{
a Seurat object with modified \code{meta.data} slot with \code{labels.num} and \code{labels.char} overwritten
}
\description{
Take a Seurat object and a labeled ident (e.g., cell_type) and produce two columns:
\itemize{
\item \code{labels.num}: vector of numbers to label inside a DimPlot
\item \code{labels.char}: vector of numbered labels to match these numbers, to use in the legend for a DimPlot
}

E.g.,

\if{html}{\out{<div class="sourceCode">}}\preformatted{labels.num   labels.char
    0         0: T cell
    1         1: B cell
    2         2: NK cell
}\if{html}{\out{</div>}}
}
\details{
As written, this will overwrite any previously stored information in the \code{labels.num} or \code{labels.char} vectors in the \code{meta.data} slot of the Seurat object

TODO: Haven't tested for edge cases with NA factor levels in the idents

TODO: See if can store this information elsewhere in the object instead of taking up room in the \code{meta.data} slot
}
\examples{
\dontrun{
so <- mk::so_label_pair(so, "cell_type")
}

}
