% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamatareprod}
\alias{squamatareprod}
\title{Squamate reptile reproductive modes}
\format{
A named vector of class \code{factor} with reproductive mode
classifications for 3962 squamates.
}
\source{
Pyron, R.A. and F.T. Burbrink. 2014. Ecology Letters 17(1): 13-21
}
\usage{
squamatareprod
}
\description{
A dataset from Pyron and Burbrink (2014) containing reproductive modes of
lizards and snakes.
}
\keyword{datasets}
