% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thm.R
\name{thm}
\alias{thm}
\alias{thm_gray}
\alias{thm_grey}
\title{Custom theme}
\usage{
thm(
  base.size = 14,
  bg = "transparent",
  title.face = "plain",
  subtitle.face = "plain",
  caption.face = "plain",
  tag.face = "bold",
  title.hjust = 0,
  subtitle.hjust = 0,
  caption.hjust = 1,
  tag.hjust = 0,
  rel.title.text = 14/14,
  rel.subtitle.text = 12/14,
  rel.caption.text = 10/14,
  rel.axis.title.text = 12/14,
  rel.axis.text = 10/14,
  rel.legend.title.text = 12/14,
  rel.legend.text = 10/14,
  rel.strip.text = 10/14,
  rel.tag.text = 16/14,
  center.titles = NULL
)

thm_gray(
  base.size = 14,
  bg = "transparent",
  title.face = "plain",
  subtitle.face = "plain",
  caption.face = "plain",
  tag.face = "bold",
  title.hjust = 0,
  subtitle.hjust = 0,
  caption.hjust = 1,
  tag.hjust = 0,
  rel.title.text = 14/14,
  rel.subtitle.text = 12/14,
  rel.caption.text = 10/14,
  rel.axis.title.text = 12/14,
  rel.axis.text = 10/14,
  rel.legend.title.text = 12/14,
  rel.legend.text = 10/14,
  rel.strip.text = 10/14,
  rel.tag.text = 16/14,
  center.titles = NULL
)

thm_grey(
  base.size = 14,
  bg = "transparent",
  title.face = "plain",
  subtitle.face = "plain",
  caption.face = "plain",
  tag.face = "bold",
  title.hjust = 0,
  subtitle.hjust = 0,
  caption.hjust = 1,
  tag.hjust = 0,
  rel.title.text = 14/14,
  rel.subtitle.text = 12/14,
  rel.caption.text = 10/14,
  rel.axis.title.text = 12/14,
  rel.axis.text = 10/14,
  rel.legend.title.text = 12/14,
  rel.legend.text = 10/14,
  rel.strip.text = 10/14,
  rel.tag.text = 16/14,
  center.titles = NULL
)
}
\arguments{
\item{base.size}{base size of text elements, also the text size of the title}

\item{bg}{color of background elements, default = "transparent". Determines background of panel, plot, legend, legend box, and strip background. For thm_gray(), it will not affect panel and strip background (which will be set gray)}

\item{title.face, subtitle.face, caption.face, tag.face}{face of main labs elements, e.g., "plain", "bold", "italic"}

\item{title.hjust, subtitle.hjust, caption.hjust, tag.hjust}{hjust of main labs elements, e.g., 0, 0.5, or 1}

\item{rel.title.text, rel.subtitle.text, rel.caption.text, rel.axis.title.text, rel.axis.text, rel.legend.title.text, rel.legend.text, rel.strip.text, rel.tag.text}{proportions for relative sizes of these elements relative to the \code{base.size}. E.g., base.size = 14 pts, then rel.subtitle.text = (12 / 14) * 14 = 12 pts}

\item{center.titles}{if TRUE, override other options and set hjust = 0 for title, subtitle, and caption}
}
\value{
A \code{theme} class object (output of \code{ggplot2::theme()})
}
\description{
Result of \code{ggplot2::theme()} to add to a ggplot with \code{+}
}
\details{
Inspired and adapted by \code{cowplot::theme_cowplot()}. Argument names and order currently in development, subject to change (except for first argument \code{base.size}).
}
\examples{
\dontrun{
# Use themes with `+`
p <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(hwy)) + ggplot2::geom_histogram()
p + mk::thm()
p + mk::thm_gray()

# Set the default theme at the top of a script
library(ggplot2)
theme_set(theme_get() + mk::thm())
theme_set(theme_get() + mk::thm(center.titles = TRUE))
}
}
