% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thm.R
\name{thm_bg}
\alias{thm_bg}
\title{Set background colors in a ggplot}
\usage{
thm_bg(
  all = NULL,
  panel = NULL,
  plot = NULL,
  legend = NULL,
  legend.box = NULL,
  strip = NULL
)
}
\arguments{
\item{all}{a color or "transparent" to apply to all elements, if NULL, need to specify one or more of the other more specific arguments}

\item{panel, plot, legend, legend.box, strip}{individually set backgrounds of these elements of the theme(), either or a color or "transparent", will invidiually override setting in \code{all}. NULL to not override anything. or if \code{!is.null(all)}, use FALSE to exclude a setting}
}
\value{
A \code{theme} class object (output of \code{ggplot2::theme()})
}
\description{
Set background of panel, plot, legend, legend box, and/or strip
}
\examples{
\dontrun{
# Example plot
set.seed(123)
df <- ggplot2::mpg |>
  dplyr::group_by(model) |>
  dplyr::summarize(hwy = mean(hwy, na.rm = TRUE)) |>
  dplyr::slice_sample(n = 4) |>
  dplyr::mutate(model2 = model)
p <- ggplot2::ggplot(data = df, mapping = ggplot2::aes(x = model, y = hwy)) +
  ggplot2::facet_wrap(
    ggplot2::vars(model), nrow = 1, scales = "free_x",
    labeller = \(x) df[, "model"] |> lapply(\(x) stringr::str_wrap(x, 10)),
    strip.position = "top"
  ) +
  ggplot2::geom_col(mapping = ggplot2::aes(fill = model)) +
  ggplot2::geom_point(mapping = ggplot2::aes(color = model2)) +
  ggplot2::labs(title = "Title", subtitle = "Subtitle", caption = "Caption")

# Original plot
p

# Transparency in all areas
p + mk::thm_bg("transparent")

# Transparency only in the overall plot background
p + mk::thm_bg(plot = "transparent")

# Transparency in all, but exclude panel and strip
  # from modification (FALSE instead of NULL)
p + mk::thm_bg("transparent", panel = FALSE, strip = FALSE)

# Transparency in all, but gray for panel and strip
p + mk::thm_bg("transparent", panel = "grey92", strip = "grey85")

# Indvidiually supply backgrounds to certain areas
p + mk::thm_bg(panel = "#2C3E50")

p + mk::thm_bg(
  panel = "maroon",
  plot = "purple",
  legend = "blue",
  legend.box = "pink",
  strip = "green"
)
}
}
