% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Waterfall plot for GSEA results}
\usage{
waterfall_plot(
  df,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  x = NULL,
  y = "NES",
  sig = c("qvalue"),
  type = c("gradient"),
  fill.limits = NULL,
  fill.breaks = NULL,
  fill.labels = NULL,
  fill.preset = NULL,
  colors = NULL,
  fill.title = NA,
  base.size = NULL,
  format = TRUE,
  .gsub.hallmark = TRUE,
  .gsub.underscore = TRUE
)
}
\arguments{
\item{df}{A dataframe of gsea results, must include columns: Description, NES, and a significance column e.g., qvalue}

\item{title, subtitle, caption}{ggplot titles, etc.}

\item{x, y}{xlab and ylab for ggplot, default NULL and "NES"}

\item{sig}{Column of \code{df} to use for significance, default = "qvalue"}

\item{type}{Type of waterfall plot, only current option is "gradient" to fill bars according to continuous significance value}

\item{fill.limits, fill.breaks, fill.labels}{Arguments to ggplot2::scale_fill_gradientn() when using \code{type = "gradient"}}

\item{fill.preset}{default NULL, or ignore the fill arguments above and use some preset values, one of: c("0.05", "0.15", "0.25")}

\item{colors, }{default NULL to use a red color scheme for \code{type = "gradient"}, or a vector of colors}

\item{fill.title}{default NA for default title, or use NULL or a custom title for the fill legend}

\item{base.size}{pass argument to mk::thm() to change default sizes of text elements in plot, default NULL to take the default value in mk::thm()}

\item{format}{TRUE (default) = consider the formatting arguments below in order, FALSE = turn all off}

\item{.gsub.hallmark}{TRUE (default) = Remove pattern ^HALLMARK_ from Description column}

\item{.gsub.underscore}{TRUE (default) = Replace pattern "_" with " " from Description column}
}
\value{
A ggplot object
}
\description{
Take df of GSEA results and return waterfall plot
}
\examples{
\dontrun{
mk::waterfall_plot(
  df = df |> dplyr::filter(qvalue < 0.05),
  sig = "qvalue",
  fill.preset = "0.05",
  title = "Hallmark Pathways",
  subtitle = "Group A vs. Group B",
  caption = "Pathways for q < 0.05"
)
}
}
